# [select] Defer mounting until typeahead is needed