--- a/docs/src/app/(private)/experiments/long-select.tsx
+++ b/docs/src/app/(private)/experiments/long-select.tsx
@@ -6,7 +6,7 @@ export default function ExampleSelect() {
   return (
     <Select.Root defaultValue="af">
       <Select.Trigger className={styles.Select}>
-        <Select.Value initial="Afghanistan" />
+        <Select.Value>Afghanistan</Select.Value>
         <Select.Icon className={styles.SelectIcon}>
           <ChevronUpDownIcon />
         </Select.Icon>