--- a/docs/src/app/(private)/experiments/scroll-area/inside-select.tsx
+++ b/docs/src/app/(private)/experiments/scroll-area/inside-select.tsx
@@ -7,7 +7,7 @@ export default function ExampleSelect() {
   return (
     <Select.Root defaultValue="item-1">
       <Select.Trigger className={styles.Select}>
-        <Select.Value placeholder="Item 1" />
+        <Select.Value initial="Item 1" />
         <Select.Icon className={styles.SelectIcon}>
           <ChevronUpDownIcon />
         </Select.Icon>