--- a/docs/src/app/(private)/experiments/toolbar/text-editor.tsx
+++ b/docs/src/app/(private)/experiments/toolbar/text-editor.tsx
@@ -141,7 +141,7 @@ export default function App() {
                 render={<Select.Trigger />}
                 className={styles.select.Select}
               >
-                <Select.Value placeholder="Sans-serif" />
+                <Select.Value initial="Sans-serif" />
                 <Select.Icon className={styles.select.SelectIcon}>
                   <ChevronUpDownIcon />
                 </Select.Icon>