--- a/docs/src/app/(public)/(content)/react/components/select/demos/hero/tailwind/index.tsx
+++ b/docs/src/app/(public)/(content)/react/components/select/demos/hero/tailwind/index.tsx
@@ -5,7 +5,7 @@ export default function ExampleSelect() {
   return (
     <Select.Root defaultValue="sans">
       <Select.Trigger className="flex h-10 min-w-36 items-center justify-between gap-3 rounded-md border border-gray-200 pr-3 pl-3.5 text-base text-gray-900 select-none hover:bg-gray-100 focus-visible:outline focus-visible:outline-2 focus-visible:-outline-offset-1 focus-visible:outline-blue-800 active:bg-gray-100 data-[popup-open]:bg-gray-100">
-        <Select.Value initial="Sans-serif" />
+        <Select.Value>Sans-serif</Select.Value>
         <Select.Icon className="flex">
           <ChevronUpDownIcon />
         </Select.Icon>