--- a/docs/src/app/(public)/(content)/react/components/toolbar/demos/hero/css-modules/index.tsx
+++ b/docs/src/app/(public)/(content)/react/components/toolbar/demos/hero/css-modules/index.tsx
@@ -38,7 +38,7 @@ export default function ExampleToolbar() {
       <Toolbar.Separator className={styles.Separator} />
       <Select.Root defaultValue="helvetica">
         <Toolbar.Button render={<Select.Trigger />} className={styles.Button}>
-          <Select.Value initial="Helvetica" />
+          <Select.Value>Helvetica</Select.Value>
           <Select.Icon>
             <ChevronUpDownIcon />
           </Select.Icon>