--- a/docs/src/app/(public)/(content)/react/components/toolbar/demos/hero/tailwind/index.tsx
+++ b/docs/src/app/(public)/(content)/react/components/toolbar/demos/hero/tailwind/index.tsx
@@ -46,7 +46,7 @@ export default function ExampleToolbar() {
           render={<Select.Trigger />}
           className="flex h-8 min-w-32 items-center justify-between rounded-sm px-[0.75rem] font-[inherit] text-sm font-medium text-gray-600 select-none hover:bg-gray-100 focus-visible:bg-none focus-visible:outline focus-visible:outline-2 focus-visible:-outline-offset-1 focus-visible:outline-blue-800 active:bg-gray-200 data-[pressed]:bg-gray-100 data-[pressed]:text-gray-900"
         >
-          <Select.Value initial="Helvetica" />
+          <Select.Value>Helvetica</Select.Value>
           <Select.Icon>
             <ChevronUpDownIcon />
           </Select.Icon>