--- a/docs/src/app/(public)/(content)/react/handbook/animation/page.mdx
+++ b/docs/src/app/(public)/(content)/react/handbook/animation/page.mdx
@@ -135,7 +135,7 @@ function App() {
   return (
     <Select.Root open={open} onOpenChange={setOpen}>
       <Select.Trigger>
-        <Select.Value />
+        <Select.Value>Value</Select.Value>
       </Select.Trigger>
       <Select.Portal>
         <Select.Positioner>