--- a/docs/src/components/Demo/DemoVariantSelector.tsx
+++ b/docs/src/components/Demo/DemoVariantSelector.tsx
@@ -126,9 +126,9 @@ export function DemoVariantSelector({
       {renderLanguageSelector && (
         <Select.Root value={selectedLocalVariant.language} onValueChange={handleLanguageChange}>
           <Select.Trigger
-            ssrFallback={
+            initial={
               currentVariantLanguages.find((item) => item.value === selectedLocalVariant.language)
-                ?.label
+                ?.label ?? ''
             }
           />
           <Select.Popup>
@@ -144,7 +144,7 @@ export function DemoVariantSelector({
       {renderVariantSelector && (
         <Select.Root value={selectedLocalVariant.name} onValueChange={handleVariantChange}>
           <Select.Trigger
-            ssrFallback={translations.variants[selectedLocalVariant.name]}
+            initial={translations.variants[selectedLocalVariant.name]}
             aria-label="Styling method"
           />
           <Select.Popup>