--- a/docs/src/components/Select.tsx
+++ b/docs/src/components/Select.tsx
@@ -6,16 +6,15 @@ import { ThickCheckIcon } from '../icons/ThickCheckIcon';
 
 export const Root = Select.Root;
 
-interface TriggerProps extends Select.Trigger.Props {
-  ssrFallback?: string;
-  initial: Select.Value.Props['initial'];
+interface TriggerProps extends Omit<Select.Trigger.Props, 'children'> {
+  children: Select.Value.Props['children'];
 }
 
-export function Trigger({ className, ssrFallback, initial, ...props }: TriggerProps) {
+export function Trigger({ className, children, ...props }: TriggerProps) {
   return (
     // Implicitly relying on <GhostButton>, keep it in sync
     <Select.Trigger data-layout="text" className="GhostButton" {...props}>
-      <Select.Value initial={initial} />
+      <Select.Value>{children}</Select.Value>
       <Select.Icon render={<ChevronDownIcon className="-ml-0.5" />} />
     </Select.Trigger>
   );