--- a/packages/react/src/select/item-indicator/SelectItemIndicator.test.tsx
+++ b/packages/react/src/select/item-indicator/SelectItemIndicator.test.tsx
@@ -11,7 +11,7 @@ describe('<Select.ItemIndicator />', () => {
       return render(
         <Select.Root open>
           <Select.Trigger>
-            <Select.Value />
+            <Select.Value initial="" />
           </Select.Trigger>
           <Select.Positioner>
             <Select.Item>{node}</Select.Item>