--- a/packages/react/src/select/root/SelectRoot.test.tsx
+++ b/packages/react/src/select/root/SelectRoot.test.tsx
@@ -32,7 +32,6 @@ describe('<Select.Root />', () => {
     render,
     triggerMouseAction: 'click',
     expectedPopupRole: 'listbox',
-    alwaysMounted: true,
   });
 
   describe('prop: defaultValue', () => {