This defers mounting the `<Select.Portal>` component until it's needed for closed typeahead (when the trigger is focused). After it opens for the first time, it stays mounted, which helps performance when opening/closing large lists.

## `Select.Value` API change
The `placeholder` prop has been changed to `children` as a plain string. This must always be specified because the items are no longer rendered upon mount, which was previously used to grab the selected item value's `.textContent` to render the value in the trigger while it was closed. However, they always needed to use `placeholder` for SSR too, and it was confusing how that API worked. The docs were using:

```jsx
<Select.Value>
  {(label) => label || ssrFallback}
</Select.Value>
```

So we could make `children: React.ReactNode` the "default/initial" label present during SSR, and now with this PR, before the select menu is opened for the first time. It will get overwritten by the selected label when it needs to after being interacted with:

```jsx
<Select.Value>
  {initialValue}
</Select.Value>
```

## Performance 

For 1,000 items: since the items don't need to mount on page load, scripting time is 35% of what it was before:

This PR: https://deploy-preview-1906--base-ui.netlify.app/experiments/select-perf
<img width="186" alt="Screenshot 2025-05-14 at 4 34 40 pm" src="https://github.com/user-attachments/assets/e4e8c933-4e16-4c6f-af1b-7bb026af7c07" />

Master: https://master--base-ui.netlify.app/experiments/select-perf
<img width="186" alt="Screenshot 2025-05-14 at 4 34 20 pm" src="https://github.com/user-attachments/assets/ce135449-cef5-4199-bfa6-18b1e85a7b92" />

However, there's a tradeoff here. When opening the Select for the first time, it's noticeably slow compared to before, since the work was already done on load when the user is less likely to notice (depending on the page/flow):

With 20x slowdown:
Master: ~1,168 ms to open first time (~40ms without slowdown)
This PR: ~2,544 ms to open first time (~136ms without slowdown)