@atomiks 2025-05-13
@romgrk the main issue here is that `<Select.Value>` uses the label of the rendered item (by default its `textContent`) after hydration. It swaps from `placeholder` to the rendered value, but this can't be deferred - so it seems that this would need to be either opt-in, _or_ require the user specify `placeholder` (or some better-named prop) for the first render