@michaldudak 2025-05-21
> However, there's a tradeoff here. When opening the Select for the first time, it's noticeably slow compared to before, since the work was already done on load when the user is less likely to notice (depending on the page/flow)

One way to solve it would be to use an intersection observer to mount only visible selects. This should help, especially on mobile devices, as their viewport is much smaller. We could also make this behavior controllable with a prop (`mountItems='eager|lazy|when-visible'`)