@romgrk 2025-05-21
> One way to solve it would be to use an intersection observer to mount only visible selects

All it takes is a couple of select with lots of items at the top of the page and the end-user will feel the delays. I don't think there's a viable solution that allows us to decide for users that we're going to dedicate CPU runtime for non-visible content.

> it's noticeably slow compared to before

I've been holding off on optimizing select items while this PR is open, but I think them being split into outer/inner components contributes to their high initial mount time. If we spend time optimizing the item component, I think we can reduce the INP for select open to acceptable levels. If it's really necessary, we could also look into virtualizing large item lists.