@atomiks 2025-05-21
I'm ok with the tradeoff of larger lists taking longer to open for now