@michaldudak 2025-05-21
Can we then see how much the current implementation can be sped up and then decide on deferring mounting strategy? We also considered using an idle callback (or timeout) to make allow the page to load and only then mount the items.