@atomiks 2025-05-21
Thinking over `Select.Value` API more: `children` as a string represents the initially selected item label during SSR. There should be no need for an extra `placeholder` prop?

Case 1: There is no way to unselect anything/no empty value

```jsx
<Select.Root defaultValue="red">
  <Select.Value>
    Red {/* SSR label */}
  </Select.Value>
  <Select.Portal>
    <Select.Positioner>
      <Select.Popup>
        <Select.Item value="red">
          Red
        </Select.Item>
      </Select.Popup>
    </Select.Positioner>
  </Select.Portal>
</Select.Root>
```

Case 2: There is an empty value but only initially

```jsx
<Select.Root>
  <Select.Value>
    Select value {/* SSR label AND placeholder, can't revert */}
  </Select.Value>
  <Select.Portal>
    <Select.Positioner>
      <Select.Popup>
        <Select.Item value="red">
          Red
        </Select.Item>
      </Select.Popup>
    </Select.Positioner>
  </Select.Portal>
</Select.Root>
```

Case 3: There is an empty value, including as an item

```jsx
<Select.Root>
  <Select.Value>
    Select value {/* SSR label */}
  </Select.Value>
  <Select.Portal>
    <Select.Positioner>
      <Select.Popup>
        <Select.Item> {/* null value by default */}
          Select value {/* Acts as the placeholder after SSR, which is the same */}
        </Select.Item>
        <Select.Item value="red">
          Red
        </Select.Item>
      </Select.Popup>
    </Select.Positioner>
  </Select.Portal>
</Select.Root>
```